moving 5 features into packages
moved profile, region, stackscript, user, and volume into their own
packages. moved a couple helpers around where they belong respective to
their package. all packages are passing tests
