+++ linode/region/datasource_test.go
@@ -1,13 +1,14 @@
-package linode
+package region_test
 
 import (
 	"fmt"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/linode/terraform-provider-linode/linode/acceptance"
 )
 
-func TestAccDataSourceLinodeRegion_basic(t *testing.T) {
+func TestAccDataSourceRegion_basic(t *testing.T) {
 	t.Parallel()
 
 	country := "us"
@@ -15,11 +16,11 @@ func TestAccDataSourceLinodeRegion_basic(t *testing.T) {
 	resourceName := "data.linode_region.foobar"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
-				Config: testDataSourceLinodeRegionBasic(regionID),
+				Config: dataSourceConfigBasic(regionID),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "country", country),
 					resource.TestCheckResourceAttr(resourceName, "id", regionID),
@@ -29,7 +30,7 @@ func TestAccDataSourceLinodeRegion_basic(t *testing.T) {
 	})
 }
 
-func testDataSourceLinodeRegionBasic(regionID string) string {
+func dataSourceConfigBasic(regionID string) string {
 	return fmt.Sprintf(`
 data "linode_region" "foobar" {
 	id = "%s"
