+++ linode/stackscript/helpers.go
@@ -1,4 +1,4 @@
-package linode
+package stackscript
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
