+++ linode/volume/datasoruce_test.go
@@ -1,24 +1,25 @@
-package linode
+package volume_test
 
 import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/linode/terraform-provider-linode/linode/acceptance"
 )
 
-func TestAccDataSourceLinodeVolume_basic(t *testing.T) {
+func TestAccDataSourceVolume_basic(t *testing.T) {
 	t.Parallel()
 
 	var volumeName = acctest.RandomWithPrefix("tf_test")
 	resourceName := "data.linode_volume.foobar"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccCheckLinodeVolumeConfigBasic(volumeName) + testDataSourceLinodeVolumeByID(),
+				Config: resourceConfigBasic(volumeName) + dataSourceConfigBasic(),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "region", "us-west"),
 					resource.TestCheckResourceAttr(resourceName, "size", "20"),
@@ -33,7 +34,7 @@ func TestAccDataSourceLinodeVolume_basic(t *testing.T) {
 	})
 }
 
-func testDataSourceLinodeVolumeByID() string {
+func dataSourceConfigBasic() string {
 	return `
 data "linode_volume" "foobar" {
 	id = "${linode_volume.foobar.id}"
